Cleanup battle-info.ts
