+++ src/utils/common.ts
@@ -2,6 +2,7 @@ import { MoneyFormat } from "#enums/money-format";
 import { Moves } from "#enums/moves";
 import i18next from "i18next";
 import { pokerogueApi } from "#app/plugins/api/pokerogue-api";
+import type { Variant } from "#app/sprites/variant";
 
 export type nil = null | undefined;
 
@@ -599,5 +600,20 @@ export function deepMergeObjects(a: Object, b: Object) {
     } else {
       a[key] = b[key];
     }
+  }
+}
+
+/** Get the localized shiny descriptor for the provided variant
+ * @param variant - The variant to get the shiny descriptor for
+ * @returns The localized shiny descriptor
+ */
+export function getShinyDescriptor(variant: Variant): string {
+  switch (variant) {
+    case 2:
+      return i18next.t("common:epicShiny");
+    case 1:
+      return i18next.t("common:shiny");
+    case 0:
+      return i18next.t("common:normal");
   }
 }
